+++ src/validation/util.js
@@ -1,0 +1,4 @@
+export function changeClass (label, oldClass, newClass) {
+  label.classList.remove(oldClass)
+  label.classList.add(newClass)
+}
